// web/src/proto/generated/DeviceInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type DeviceInfo = { 
/**
 * Device public key (base64url).
 */
public: string, name: string, client: string, paired_at: number, last_seen?: number, permissions: Array<string>, 
/**
 * True for the device making the request.
 */
current: boolean, };